Added bugs in the selected card effects.
Smithy – Changed the loop less than (“<”) condition to less than or equal to (“<=”); this bug would make the effect the card has to draw four cards versus three cards. Additionally, the smithy card will be sent to the trash pile versus the players discard pile since the parameter in the discardCard function was changed from a 0 to a 1.

Adventurer - Changed if statement by removing the option to check if the copper card is a treasure card. This bug will not add copper cards to the players hand, only silver or gold.

Salvager – Changed the discard index inside the if statement so if the play chooses to trash their card, it will trash their hand position card versus the card those chose to trash.

Steward – Changed the conditional statement to gain to coins to the input of choice 2 versus choice1. This might affect whether this option will be selected or not depending on the value of choice2. Additionally, both cards will only be discarded and not trashed if the trash option is selected since, the values of 0 are given versus 1, which indicated the cards should be trashed versus discarded.